Preserve OTLP gauge and counter precision
Carry OTLP gauge and sum datapoints as doubles through conversion and ClickHouse WAL rows. Keep native sketch narrowing explicit at its approximate-storage boundary and verify fractional values survive JSON serialization.

+++ okapi-ingester/src/main/java/org/okapi/metrics/ch/ChMetricsWalConsumer.java
@@ -53,7 +53,7 @@ public class ChMetricsWalConsumer {
                 .tags(req.getTags())
                 .unit(req.getUnit())
                 .timestamp(ts)
-                .value(req.getGauge().getValue().get(i))
+                .value(req.getGauge().getValue().get(i).doubleValue())
                 .build();
         gaugeSamples.add(gson.toJson(sample));
         metaRows.add(metricMetadataRow(req, METRIC_TYPE.GAUGE, null, ts, ts));
